Port tests from expresso to mocha. Fixes #11
  Expresso is no longer maintained.
