+++ package.json
@@ -1,37 +1,39 @@
 {
-    "name" : "hat",
-    "version" : "0.0.3",
-    "description" : "generate random IDs and avoid collisions",
-    "main" : "index.js",
-    "directories" : {
-        "lib" : ".",
-        "example" : "example",
-        "test" : "test"
-    },
-    "devDependencies" : {
-        "expresso" : "0.7.x"
-    },
-    "scripts" : {
-        "test" : "expresso"
-    },
-    "repository" : {
-        "type" : "git",
-        "url" : "http://github.com/substack/node-hat.git"
-    },
-    "keywords" : [
-        "id",
-        "uid",
-        "uuid",
-        "random",
-        "hat",
-        "rack",
-        "unique"
-    ],
-    "author" : {
-        "name" : "James Halliday",
-        "email" : "mail@substack.net",
-        "url" : "http://substack.net"
-    },
-    "license" : "MIT/X11",
-    "engine" : { "node" : ">=0.4" }
+  "name": "hat",
+  "version": "0.0.3",
+  "description": "generate random IDs and avoid collisions",
+  "main": "index.js",
+  "directories": {
+    "lib": ".",
+    "example": "example",
+    "test": "test"
+  },
+  "devDependencies": {
+    "mocha": "^2.2.5"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "http://github.com/substack/node-hat.git"
+  },
+  "keywords": [
+    "id",
+    "uid",
+    "uuid",
+    "random",
+    "hat",
+    "rack",
+    "unique"
+  ],
+  "author": {
+    "name": "James Halliday",
+    "email": "mail@substack.net",
+    "url": "http://substack.net"
+  },
+  "license": "MIT/X11",
+  "engine": {
+    "node": ">=0.4"
+  }
 }
